optimizes search tab
