+++ lib/src/pages/search/search_bloc.dart
@@ -1,0 +1,6 @@
+import 'package:podcast/src/bloc/bloc_base.dart';
+
+class SearchBloc extends BlocBase {
+  @override
+  void dispose() {}
+}
